requirements.txt
azure-search-documents
openai
python-dotenv
pandas
langchain
tiktoken
langchain-community
langchain-openai